Separating out protobuf conversions to make it more manageble
