+++ Source/Events/Relativity/Protobuf/Conversion/GeneralExtensions.cs
@@ -1,0 +1,166 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Dolittle. All rights reserved.
+ *  Licensed under the MIT License. See LICENSE in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+using System;
+using System.Collections.Generic;
+using System.IO;
+using Dolittle.Applications;
+using Dolittle.Artifacts;
+using Dolittle.Collections;
+using Dolittle.Concepts;
+using Dolittle.PropertyBags;
+using Dolittle.Reflection;
+using Dolittle.Serialization.Protobuf;
+using Dolittle.Time;
+using Google.Protobuf;
+using Google.Protobuf.Collections;
+using System.Globalization;
+
+namespace Dolittle.Runtime.Events.Relativity.Protobuf.Conversion
+{
+    /// <summary>
+    /// Extension methods for working with conversion related to <see cref="Dolittle.Runtime.Events.Store.CommittedEventStream"/>
+    /// </summary>
+    public static class GeneralExtensions
+    {
+        /// <summary>
+        /// Convert from a <see cref="System.Guid"/> to a <see cref="System.Protobuf.guid"/>
+        /// </summary>
+        /// <param name="guid"><see cref="System.Guid"/> to convert</param>
+        /// <returns>A <see cref="System.Protobuf.guid"/></returns>
+        public static System.Protobuf.guid ToProtobuf(this System.Guid guid)
+        {
+            var protobufGuid = new System.Protobuf.guid();
+            protobufGuid.Value = ByteString.CopyFrom(guid.ToByteArray());
+            return protobufGuid;
+        }
+
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="guid"></param>
+        /// <returns></returns>
+        public static ByteString ToByteString(this ConceptAs<Guid> guid)
+        {
+            return ByteString.CopyFrom(guid.Value.ToByteArray());
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="concept"></param>
+        /// <returns></returns>
+        public static System.Protobuf.guid ToProtobuf(this ConceptAs<Guid> concept)
+        {
+            return concept.Value.ToProtobuf();
+        }
+
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="byteString"></param>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        public static T ToGuidConcept<T>(this ByteString byteString) where T : class
+        {
+            return ConceptFactory.CreateConceptInstance(typeof(T), new Guid(byteString.ToByteArray())) as T;
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="guid"></param>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        public static T ToConcept<T>(this System.Protobuf.guid guid) where T : ConceptAs<Guid>
+        {
+            return ConceptFactory.CreateConceptInstance(typeof(T), new Guid(guid.Value.ToByteArray())) as T;
+        }
+
+
+        /// <summary>
+        /// Get the protobuf <see cref="Types"/> representation of the <see cref="object">instances type</see>
+        /// </summary>
+        /// <param name="obj"><see cref="object">instance</see> to get from</param>
+        /// <returns><see cref="Types">Protobuf type</see></returns>
+        public static Types GetProtobufType(this object obj)
+        {
+            var type = obj.GetType();
+            if( type == typeof(string)) return Types.String;
+            if( type == typeof(int)) return Types.Int32;
+            if( type == typeof(long)) return Types.Int64;
+            if( type == typeof(uint)) return Types.UInt32;
+            if( type == typeof(long)) return Types.UInt64;
+            if( type == typeof(Int32)) return Types.Int32;
+            if( type == typeof(Int64)) return Types.Int64;
+            if( type == typeof(UInt32)) return Types.UInt32;
+            if( type == typeof(UInt64)) return Types.UInt64;
+            if( type == typeof(float)) return Types.Float;
+            if( type == typeof(double)) return Types.Double;
+            if( type == typeof(bool)) return Types.Boolean;
+            if( type == typeof(DateTime)) return Types.DateTime;
+            if( type == typeof(DateTimeOffset)) return Types.DateTimeOffset;
+            if( type == typeof(Guid)) return Types.Guid;
+
+            return Types.Unknown;
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="type"></param>
+        /// <param name="obj"></param>
+        /// <param name="stream"></param>
+        public static void WriteWithTypeTo(this object obj, Types type, CodedOutputStream stream)
+        {
+            switch( type )
+            {
+                case Types.String: stream.WriteString(obj as string); break;
+                case Types.Int32: stream.WriteInt32((int)obj); break;
+                case Types.Int64: stream.WriteInt64((Int64)obj); break;
+                case Types.UInt32: stream.WriteUInt32((uint)obj); break;
+                case Types.UInt64: stream.WriteUInt64((UInt64)obj); break;
+                case Types.Float: stream.WriteFloat((float)obj); break;
+                case Types.Double: stream.WriteDouble((double)obj); break;
+                case Types.Boolean: stream.WriteBool((bool)obj); break;
+                case Types.DateTime: stream.WriteInt64((Int64)((DateTime)obj).ToUnixTimeMilliseconds()); break;
+                case Types.DateTimeOffset: stream.WriteInt64((Int64)((DateTimeOffset)obj).ToUnixTimeMilliseconds()); break;
+                case Types.Guid: stream.WriteBytes(ByteString.CopyFrom(((Guid)obj).ToByteArray())); break;
+            }
+        }
+
+        /// <summary>
+        /// Read value from <see cref="System.Protobuf.Object"/>
+        /// </summary>
+        /// <param name="obj"><see cref="System.Protobuf.Object"/> to read from</param>
+        /// <returns>Value in the correct type - null if not capable of converting</returns>
+        public static object ConvertToCLR(this System.Protobuf.Object obj)
+        {
+            var type = (Types)obj.Type;
+            object value = null;
+
+            using( var stream = new CodedInputStream(obj.Content.ToByteArray()))
+            {
+                switch( type )
+                {
+                    case Types.String: value = stream.ReadString(); break;
+                    case Types.Int32: value = stream.ReadInt32(); break;
+                    case Types.Int64: value = stream.ReadInt64(); break;
+                    case Types.UInt32: value = stream.ReadUInt32(); break;
+                    case Types.UInt64: value = stream.ReadUInt64(); break;
+                    case Types.Float: value = stream.ReadFloat(); break;
+                    case Types.Double: value = stream.ReadDouble(); break;
+                    case Types.Boolean: value = stream.ReadBool(); break;
+                    case Types.DateTime: value = DateTime.FromFileTime(stream.ReadInt64()); break;
+                    case Types.DateTimeOffset: value = DateTimeOffset.FromFileTime(stream.ReadInt64()); break;
+                    case Types.Guid: value = new Guid(stream.ReadBytes().ToByteArray()); break;
+                }
+            }
+
+            return value;
+        }
+    }
+}
